\name{eolsdng-package}
\alias{eolsdng-package}
\alias{eolsdng}
\docType{package}
\title{
Read and Visualize EOL sounding data
}
\description{
This package contains functions for reading EOL sounding files
    and plotting the soundings.
}
\details{
\tabular{ll}{
Package: \tab eolsdng\cr
Version: \tab 2.4-268\cr
Depends: \tab lattice,maps, eolts, isfs\cr
License: \tab BSD_2_clause + file LICENSE\cr
Date: \tab 2014-05-01\cr
Packaged: \tab 2014-05-20 19:33:44 UTC; maclean\cr
Built: \tab R 3.0.2; ; 2014-05-20 19:33:45 UTC; unix\cr
}

Index:
\preformatted{
eolsdng-package         Read and Visualize EOL sounding data ~~ package
                        title ~~
plotScontour            Plot Contours of EOL Sounding Profiles
readSoundings           Function to read one or more soundings.
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Liangying Zhang
Maintainer: Gordon Maclean <maclean@ucar.edu>
}
\references{
}
\keyword{ package }
\seealso{
\code{\link[eolts:eolts-package]{eolts}}
\code{\link[isfs:isfs-package]{isfs}}
}
\examples{
}
